evn_django_core/requirements.txt
django==4.0.5
djangorestframework
pyyaml
requests
django-cors-headers
cx_oracle